Separate cogs into a file per activity (#56)
* Update dependencies & fix the outdated lock file format

* Fix invalid reminder search predicate arguments

* Allow "logging_message" argument of send_error() to be of type Exception

* Allow non-inducted users to use the "/makemember" command

* Implement "/ensure-members-inducted" command

* Move send_error() method to cogs._utils

* Allow autofix of ruff import rules

* Load all individual cogs from the cogs module __init__.py file

* Move commands to individual cog files

* Move ensure_members_inducted command to individual cog file

* Move events to individual cogs files

* Move tasks to individual cogs files

* Fix invalid validation after member name autocomplete on "/induct" command

* Rename welcome channel shortcut accessor to "rules_channel"

* Fix ruff linting errors

* Fix DISCORD_LOG_CHANNEL_WEBHOOK_URL validator

* Fix matplotlib typing errors

* Fix invalid path references when testing argument parsing using subprocess execution

* Move ClearRemindersBacklogTaskCog class to remind_me.py cog

* Move EnsureMembersInductedCommandCog cog to induct.py cog

* Fix incorrect cog imports after moving cog classes to other files

* Add docstrings

* Remove incorrectly added empty files after merge

* Fix unsorted import block
